mod supp

+++ php/Classes/Manager/UserManager.php
@@ -229,6 +229,8 @@ class UserManager{
      * @param $id
      */
     public function delUser($id){
+        $user = $this->getUserById($id);
+        if ($user->)
         $stmt = $this->db->prepare("DELETE FROM user WHERE id = :id");
         $stmt->bindValue(':id',$id);
         $stmt->execute();
